Refactor task structure and further documentation.
