+++ src/db/traits/mod.rs
@@ -7,6 +7,7 @@ pub mod object;
 pub mod permissions;
 pub mod relations;
 pub mod task;
+pub mod task_import;
 pub mod user;
 
 #[allow(unused_imports)]
