+++ src/tasks/mod.rs
@@ -1,0 +1,12 @@
+mod execution;
+mod helpers;
+mod planning;
+mod resolution;
+mod types;
+mod worker;
+
+pub use helpers::request_hash;
+pub use worker::{ensure_task_worker_running, kick_task_worker};
+
+#[cfg(test)]
+mod tests;
