+++ src/utilities/mod.rs
@@ -5,7 +5,6 @@ pub mod iam;
 pub mod init;
 pub mod reporting;
 pub mod response;
-pub mod tasks;
 
 pub fn is_valid_log_level(level: &str) -> bool {
     matches!(level, "error" | "warn" | "info" | "debug" | "trace")
